requirements/test.txt
-r prod.txt
django-debug-toolbar==4.3.0
flake8==7.1.1
flake8-commas==4.0.0
flake8-dunder-all==0.4.1
flake8-import-order==0.18.2
flake8-quotes==3.4.0
parametrize==0.1.1
pep8-naming==0.14.1
